requirements.txt
streamlit
fastapi
uvicorn
transformers
torch
pytesseract
Pillow
numpy
PyMuPDF
requests
google-generativeai